Implement the nuChic code as a Package
Allow the code to run completely with Quasielastic and cascade together
through a common interface with absl-py. This required an update of the
setup.py file to be able to build it all correctly.

Also address the data files to be stored with the code and found based
off the installation location.

Implement some basic logging and flags for the code to run with.