Implement a histogram class.
Create an initial histogram class for storing observables in. This
contains basic information about the binning before the code runs. This
allows the user to fill the histograms as the program runs instead of
trying to store all the information in arrays which get converted into
histograms at the end. The class contains the ability to calculate the
integral of the histogram, and stores both the weights and weights^2 to
allow for other statistical calculations which have yet to be
implemented.

Additionally, unit tests for the class are also implemented for all
functions except for the saving and plotting routines. I am unsure how
to properly test these without needing to run I/O or the matplotlib
interface which will break the idea of unit testing.

TODO:
    - Write tests for saving and plotting
    - Implement other statistics for the histogram (such as mean,
    variance, etc.)
    - Write a parser that will load this from a text file (Maybe?)