Implement angular interactions in cascade.
Use the GEANT data for both cross-sections and angular distributions for
the cross-sections as a function of center of mass energy. In the case
that the center of mass energy falls outside the GEANT range, fall back
on the previous method of calculating cross-section. In the case of
throwing a random number that is smaller than the cross-section for 0.5
degrees, linearly interpolate between 0 and 0.5 for the angle.

Additional changes include warnings about when the above happen, and
some other formatting and other clean-up of the code.

TODO:
    - Additional refactoring
    - Hooking up histogramming correctly
    - Other utility functions
    - etc.