Implement weighted event generation
The default vegas code found on pypi will not generate an exactly
requested number of events. Here, the vegas class is subclassed to allow
for the generation of a user requested number of events, along with the
weights of the events, and an estimate of the total integral.

Additionally, a preliminary attempt for generating unweighted events is
partially implemented. Here, a rough estimate of the maximum weight that
would be generated for a fixed number of events is calculated. Whether
or not this is the best way of doing it needs to be examined.