Refactor code to conform with PEP8 and pylint
Continue to refactor the code to conform to standards. Most of the code
has now been finished with the refactoring. Need to address the main
function, which was a placeholder for testing to begin with anyways. The
tests are all passing now with the new formatting.

To help improve finding the correct version of files, the code has been
updated to use relative paths instead of importing from the nuchic
package.