Develop a basic integration over a delta function
Attempt to figure out how to integrate the four variable
functional form efficiently. Currently, this does not work to give an
accurate and stable result. Right now, a gaussian with a width is used
to approximate a delta function. Tuning the width is required for
stability, or a new method needs to be developed.

Other minor changes include a progress bar to show events being
generated instead of printing out every n events, and other changes to
get the main code to conform to PEP8 and pylint standards.