Implement an input parser
In order to allow the user to define all the runtime parameters in a
file an input parser was needed. Here, we use a YAML based syntax along
with pyYAML to achieve this task.

Currently, only basic input options are available and are not used in
the main code with the exception of the number of events. Ideally, the
user would be able to define all possible settings, or omit any ones to
use the defaults. To this end, the default settings will still need to
be implemented, along with checking that the user does not put an
invalid setting in the run card.

Some additional thoughts are should we allow the user to define
histograms here. If so, how do we go about capturing a functional form
in order to properly fill the histogram?

Finally, I think it would be ideal if the command line arguments would
override the input file. This still needs to be implemented.

TODO:
    - Additional input variables
    - Hook up the input to where they are actually needed
    - Override if given a command line argument
    - Take a command line to define what the run card to be ran should
    be

Optional and Questions:
    - Allow user histograms?
    - Capture histogram fill functions somehow