Enforce PEP8 and pylint standards
The code is now almost completely compliant to the PEP8 and pylint
standards. There are some names of functions and variables that could be
improved. The only real culprit left is cascade.py which needs to be
looked into in great detail for refactoring and other improvements.
Therefore, this is left to a future commit.