Squash some bugs
Found bugs in the implementation of the folding interface. Shift is
still not correct, but this fix may require a bigger rewrite.
Additionally, remove bugs that remained when the \cos\theta integration
was removed.

Implement plotting of the data with the correct histograms by ensuring
the plots get the widths of the bins divided by the value. Maybe this
should be an option?