Correct units and potential contribution
Adjust the calculation to subtract the contribution of the potential
from the momentum instead of adding it. This is the correct way since
the potential is positive, but we need an attractive potential.

Additionally, convert the calculation to appropriately give the
cross-section instead of multiplying it by an additional factor of 1000.
This then required an adjustment in the binning of the cross-section as
a function of omega. Instead of filling it with MeV, it must be filled
with GeV in order for our calculation to be consistent with the data.