Implement adjustable max steps in Cascade
In the cascade code, it would traditionally run for a fixed number of
steps set by default. This was not an issue in the case of electron
events, since it had an additional stopping condition when all nucleons
were outside the nucleus. However, this did not work with the
calculation of the pN cross-section due to the fact that we started
outside the nucleus. By setting the max steps to the time it takes to
transverse the nucleus, this greatly speed up the code.