Ignore capitalization in string settings
The settings that took string variables depended on ensuring the
capitalization structure matched that of the code. This corrects it for
the three current settings. Maybe there is a way to ensure this works
for future settings by adding a flag when reading in the strings, and
making them all either be all capital or all lower case?