Implement utility objects
The Random.hh contains a class for generating random numbers from:
pcg-random.org/posts/ease-of-use-without-loss-of-power.html. In this
post, the author discusses issues with random number generation in
C++11. This code attempts to mitigate this.

Additionally, other utilities are implemented in the Utilities.hh and
Utilities.cc to reduce the number of repeated code.

TODO:
- Figure out how to have only one random class object throughout the
  code, including on the python side
- Determine if any other utilities are needed