Begin documentation of the code
Use sphinx, breathe, and exhale to put together documentation for both
the python and C++ code. The C++ code is documented using Doxygen style
commetns, while the python is documented using Sphinx.

As of right now, the ThreeVector class is documented briefly. This may
need to be extended and updated to have more details.