Add function MeanFreePath
The new function MeanFreePath evolves a kicked particle until its
first interaction with another particle, accumulating the total
distance travelled along the way. Currently, MeanFreePath is
simply a stand-alone function with an implementation very similar
to Cascade::Operator. In the future, it may be desirable from the
perspective of unit testing to modify Cascade::Operator to measure
distanceTraveled directly.