Implement setuptools_scm to maintain version
This also adds in a printout of the version into the log file for
debugging purposes, and for helping users when the program is eventually
made public. The information will include the commits and if the code is
dirty. This will allow the developers to help track done the bugs and to
point out to the user if their changes may have broken the code.