Address clang compilation bug
The clang compiler is agressive with the -= and /= operators. This
causes the definition of these operators using py::self to raise
self-assign-overloaded warnings. This is discussed in some detail here:

https://bugs.llvm.org/show_bug.cgi?id=43124

However, it seems to not be included into clang yet. Therefore, a
workaround is currently implemented to avoid this issue.