Correct mean free path calculation
The mean free path needs to be calculated in a constant density.
Therefore, choosing a random particle biases the result. This is due to
the fact that the nucleons on the edge see a different density then
those on the interior. Additionally, the expected result calculation has
been added to the plot and numerically to compare the results.