Implement the Transparency run mode
The transparency run mode has been added. This is based on the method in
the NuWro paper. The algorithm is as follows:

1. Randomly select a nucleon inside the nucleus
2. Give the nucleon a kick
3. Propagate the nucleon
4. If it interacts, stop the calculation and increment the interaction
counter.
5. Otherwise, continue until it escapes.
6. The transparency is the number of events in which the nucleon escapes
with no interaction divided by the total number of events.

This result has been briefly compared at a few points to the NuWro
result, and seems consistent at low energies. However, it appears to
disagree at high energies which may be due to the fact that pions are
currently not implemented in this algorithm.