Refactor Cascade
Modify the evolution of the nucleus such that it actually is changed.
This involved checking that the configurations are correctly generated
only on the first call. This could have issues for configuration
functions that do not always return the same length. Additionally, now
the different evolution classes do not return anything, but modify the
past in nucleus.