Implement basic ParticleInfo classes
These classes include: a PID class, a ParticleInfoEntry class, and a
ParticleInfo class. The PID class implements a basic expandable enum
class with some common PIDs predefined as static constant expressions.
The ParticleInfoEntry class defines an entry that gets loaded into a
ParticleDatabase to store the properties of all the particles to be used
throughout the program. Finally, the ParticleInfo class holds a specific
instance of a ParticleInfoEntry along with whether it is a particle or
an anti-particle. This allows global updates to the database to
propagate through all the particles created of that type.