Hotfix: Resolve issue with Segfaulting
The segfault was due to the Particle Database not properly being loaded
in before trying to create a particle. This caused the code to attempt
to search an undefined map. This has now been resolved, and the code
will work. However, another bug with the GeantInteractions has been
found. This will be resolved in the subsequent commit.

TODO:
    - Resolve issue with GeantInteractions