Resolve GeantInteractions loading bug
The error that is thrown from the Brent root finding code was changed
from runtime_error to domain_error. This was done, since it makes more
sense to say it is a domain_error if there is no root in the range. This
makes the error clearer and less generic. However, this change was not
propagated to the loading of the Geant Interaction data to linearly
interpolate in the small angle region.