Resolve transparency_issue
The issue was caused by a bug that was introduced when refactoring. The
kicked index list was not reset each time. This resulted in the code
propagating the incorrect nucleon and thus the internal_test particle
would never escape. Additionally, when the particle escaped, the status
code was not updated to signify this. This would also make the analysis
code at the end note that this particle interacted.