Finish Metropolis interaction model
Follow the Metropolis algorithm for the cascade. This introduces pion
interactions in a fairly straightforward way. The code still needs to be
tested. The cascade interface was also modified to allow for this
interaction model to connect correctly. This also resulted in needing to
change the old interaction models to match this new interface.