Interface Fortran to basic physics classes
In order for the end-user to be able to extend the generator in
Fortran, some basic interface needed to be developed. This is a start
in that direction. Here we implement the ability to interface with:
1. logging framework
2. particle information database
3. four vectors
4. three vectors
5. particle class

The next steps include figuring out if it is possible to extend the
Interactions class in a reasonable way, and to develop a Fortran
factory for this. The issue is figuring out registration to the
factory, and implementing dynamic library loading.