Separate fortran interface into two libraries
The separation puts the fortran files in one library, and the c++ files
into another. This allows the appropriate compiler flags to be sent to
each file. This is more of an issue with how CMake handles compiler
flags. For example, it is not possible to pass different standards based
on the language.