Refactor HardScattering and add Pions
Reorganizing the fortran folder was necessary for maintainability. It
was becoming to cluttered. Additionally, the fortran interface to the
hard scattering has been setup to be easily extensible. This allows
easier maintainability for each additional process, and makes the
process of adding a new hard scattering cross section more
straightforward.