Implement batch running, hard vs event-level cuts
Event generation can now be run using Vegas in batches, which allows the
program to keep running until a threshold number of events has been
passed. This functionality is especially useful when stringent cuts,
which can dramatically reduce the number of events, are imposed.
This commit also makes a distinction between "hard cuts" (which occur
before the cascade) and "event cuts" which occur after the cascade.
For the latter, an event is accepted after a single particle passes all
appliciable cuts. For instance, a single proton might be required to
have energy and angle in (E+/-dE, phi+/-dphi).