Fix bug with applying event cuts
Previously, EventGen::Calculate would return zero if it failed the
event cut. This behavior was incorrect, since it caused Vegas to
mistakenly interpret the integrand as zero in this region and thus
incorrectly adapt its underlying grid. In fact, we just want to skip
saving such events, while still returning the correct weight.