Refactor Cascade Interaction Models
When wanting to consider multiple different interaction models for
different initial state particles, the previous method was not able to
properly handle this. Therefore, a CascadeInteraction class was created
to hold all the allowed interactions during the cascade. Components are
then added to it based on the run card from the user. Currently, there
is no error checking to make sure that all the different models are
consistent and do not describe the same set of interactions.
Furthermore, the current framework needs some work to handle the
difference between \pi N -> \pi N and \pi N -> \Delta.

Hopefully, this refactoring will allow to more easily implement these
changes, and make the code modular enough to swap in and out different
models on the side of the user.