Implement basic symplectic integrator
Evolving the nucleons within a potential while maintaining the energy of
the nucleons, requires a symplectic integration technique. Since the
Hamiltonian is unseparable (due to the potential being a function of
both the momentum and the position of the nucleon), we need a method to
handle this situation. We therefore implement the phase-space extension
technique outlined in Phys. Rev. E. 94 (043303).

The code seems to work well for a test case, but the 4th and 6th order
implementations through the use of "the triple jump" do not seem to make
a difference. This needs to be investigated further.