WIP: Implement new VEGAS routines
In the end, we want to generate a fixed number of events, and to also
include unweighting. To make this work we need to rework the VEGAS
routines and related classes. In addition, adding in the multi-channel
integration will require a reworking of the VEGAS to have better
communication between the two.