Improve integration routines
Update the Vegas integration routines and implement a multichannel
integrator. Both routines now have an optimize routine that will take
the input function, and calculate the desired integral until a desired
precision is reached.

Additionally, a method of serializing the current state into a YAML
format was created to allow for saving and loading of the results. This
should reduce the time needed when rerunning with the same parameters.

The only thing that is currently missing from this implementation is the
ability to save the Mapper that was used to transform the variables from
the unit hypercube into each channel. This requires a means to write out
the type of the channel, and then load that instance correctly. The plan
is to do this with a self-registering factory based on the blog
https://www.jibbow.com/posts/cpp-header-only-self-registering-types/.