WIP: Continue testing and benchmarking autodiff
Created a dummy fuction to see if the speed difference holds for other
functions. Autodiff seems to be faster for other functions, so there is
something wrong with evaluating the potential. Also, switching from
autodiff::dual to autodiff::real drastically improved the speed of the
code. The other thing to attempt to check is to see if it is possible to
rewrite how the potential returns things to see if autodiff can give all
four derivatives at once.