Refactor cut classes
Reorganize the cuts to be a set of subclasses, with a driver class that
holds all the cuts and evaluates them efficiently.

This also made it easier to implement the two particle cuts, and any
future needed generator level cuts. Furthermore, adding in new cuts is
straightforward, and just requires the use of one macro function, and
for the user to define the observable to be cut on.