Correct issues with Channel updates
Previously, if a channel had no events in a given iteration, the
histogram would still be updated. This would result in always generating
a zero for the random number. To resolve this, only if all the data is
not zero is the channel adapted.

Additionally, some debugging information was added to help trace future
issues that may arise in the AdaptiveMap