Include potential propagation
Allow the user to select the choices for in-medium effects to include,
along with the ability for the user to choose whether to propagate
according to the potential or not. Currently, this is only implemented
completely for the MeanFreePath_NuWro function, and partially for the
NuWro cascade. It still needs to be implemented for the impact parameter
cascade. In addition, the cascade code needs to be cleaned up and
refactored for better usability.