Refactor Potential classes
Remove the need of autodiff, since it does not seem to improve the speed
of the code. Also, move most of the details into the .cc file instead of
leaving them in the header.

Implements a factory class to setup the potentials to allow the user to
choose the potential from the configuration file.