Implement Weyl basis and correct contractions
The old fortran code used the Dirac basis for the spinors, while the
Sherpa code used the Weyl basis for the spinors. This caused issues with
contracting over spins (which was previously also done in the wrong
order). This resolves both of the issues. The code now produces the
correct results.