Update README.md and modify code to match README
As the README.md was updated to reflect the code, some
issues with the setup of the code were found. These were corrected.
Additionally, during the writing of the README.md, I noticed that
things could be made more clear and easier for the user by making
changes to the code. These changes are also included here.