Implement basic plugin loader
The plugin loader can load any library with AchillesPlugin contained
in the name. Currently, hooking the plugin into the system via cmake
requires some work. It should be fairly easy to develop a simple
python script that generates the basic needed code. This would
involve at a minimum setting up the cmake to include finding all
the needed Achilles libraries during configuration if the code is
standalone. It might be easier to not use CMake and just use python
to write a simple Makefile or compile directly (similar to Rivet).