Refactor memory management of nucleus
Running Achilles with an AddressSanitizer or Valgrind revealed
a serious indirect memory leak. This was caused by a circular
memory dependence between the potential and the nucleus classes.
This has been resolved by ensuring that the nucleus class owns the
potential (has a unique_ptr to it), while the potential class just
gets a pointer to the nucleus when it is needed. This required
a major rewrite, but has resolved the memory issues.