Implement handling multiple nuclei
Multiple nuclei handling has been added to the process grouping on top of the leptonic final state and the nuclear model.
This however required the additional special handling of hydrogen to allow for better simulations of MiniBooNE. The implementation
of the hydrogen handling is not ideal, but further improvements would require a significant reworking of the code. These changes
are left to a future commit when more details about the optimal approach have been worked out.