Implement flag for Sherpa with UFO2.0
Create a separation for versions of Sherpa with UFO2.0 implemented or
not. I still need to add in a check for Sherpa to see if it has UFO2.0
implemented or if it is still using 1.0.

Additionally, I need to prevent the Sherpa backend if not using UFO2.0,
and only allow the BSM backend to interface with Sherpa in this case.