feat: Status updates of particles in history
The particles in the history would not have their statuses updated if
they were changed during the cascade. This change enables that to be
correctly handled and connecting the different nodes of the graph. The
major issue with this approach is that it is very cumbersome. It would
be nice if this could be handled using std::reference_wrapper around the
particles to ensure that the history just has references to the
particles. However, the first attempt at this failed, since it could not
handle a resizing of the event.Hadrons() vector. I think it would be
worthwhile to revisit this in the near future.