feat: Implement and update rules and handling
Develop a setup for providing a means for conditional checking on
cascade interactions (e.g. ensure that if `DeltaInteraction` is set, then
the `NucleonNucleon` interaction uses `Propagate` for the
`ResonanceMode`, and the opposite for all the modes without propagating
resonances).

Develop a setup for checking that parameters fall within a valid range
(e.g. NEvents > 1, 0 < ExpSud < 2, etc.)

TODO: Implement a validation to ensure same named parameters are
consistent (ExpSud in NN and ND interaction modes)?