Settings validation (#120)
* feat: Implement global required settings check

* feat: Add basic rules file

* feat: Implement and update rules and handling

Develop a setup for providing a means for conditional checking on
cascade interactions (e.g. ensure that if `DeltaInteraction` is set, then
the `NucleonNucleon` interaction uses `Propagate` for the
`ResonanceMode`, and the opposite for all the modes without propagating
resonances).

Develop a setup for checking that parameters fall within a valid range
(e.g. NEvents > 1, 0 < ExpSud < 2, etc.)

TODO: Implement a validation to ensure same named parameters are
consistent (ExpSud in NN and ND interaction modes)?

* fix: Handle optional parameters

* fix: Remove unused error type

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>