Implement Batch Mode
New commandline argument "-b": if present, the generator progress is printed using spdlog::info instead of fmt::print (this allows it to actually show up in the log file).

Omitting -b restores the previous functionality: the progress takes up only 1 line in the command shell which continuously updates, and progress is not printed to the log file.